Move inspection function to inspector file
